app/requirements.txt
alembic==1.11.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==4.0.0
flask-session==0.5.0
Flask-SSLify==0.1.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
